requirements.txt
fastapi
uvicorn
numpy
pandas
faiss-cpu
scikit-learn
requests
pydantic-settings
pathlib
pyarrow
matplotlib